This shamelessly copies out of Pkg all of the bits required to actually download artifacts, etc. The motivation for this stdlib is to reduce the loading / dependency cost for downloading artifacts, esp. via LazyArtifacts.

With this PR (+ [LazyArtifacts.patch](https://github.com/user-attachments/files/16979700/LazyArtifacts.patch)):
```console
$ ./julia -e '@time using MicroMamba'
  0.179380 seconds (188.20 k allocations: 10.800 MiB)
```

versus master:
```console
$ ./julia -e '@time using MicroMamba'
  1.857587 seconds (932.05 k allocations: 56.118 MiB, 6.27% gc time, 13.45% compilation time: 3% of which was recompilation)
```

#### Detailed timings
PR (+ [LazyArtifacts.patch](https://github.com/user-attachments/files/16979700/LazyArtifacts.patch)):
```julia
$ ./julia -e 'using InteractiveUtils; @time_imports using MicroMamba'
      0.8 ms  Printf
     42.9 ms  Dates
      0.4 ms  Scratch
      0.5 ms  TOML
               ┌ 0.0 ms NetworkOptions.__init__()
      3.8 ms  NetworkOptions
               ┌ 0.3 ms MbedTLS_jll.__init__()
      4.7 ms  MbedTLS_jll
               ┌ 0.1 ms LibSSH2_jll.__init__()
      3.7 ms  LibSSH2_jll
               ┌ 0.2 ms LibGit2_jll.__init__()
      5.9 ms  LibGit2_jll
     21.2 ms  LibGit2
     22.7 ms  ArgTools
      1.8 ms  Tar
               ┌ 0.2 ms nghttp2_jll.__init__()
      5.2 ms  nghttp2_jll
               ┌ 0.2 ms LibCURL_jll.__init__()
      6.6 ms  LibCURL_jll
               ┌ 0.0 ms MozillaCACerts_jll.__init__()
      5.3 ms  MozillaCACerts_jll
               ┌ 0.0 ms LibCURL.__init__()
      1.8 ms  LibCURL
               ┌ 0.2 ms Downloads.Curl.__init__()
     59.5 ms  Downloads
               ┌ 0.2 ms p7zip_jll.__init__()
      9.1 ms  p7zip_jll
      2.4 ms  ArtifactDownloads
     22.9 ms  LazyArtifacts
     19.6 ms  Preferences
      0.6 ms  JLLWrappers
               ┌ 0.2 ms micromamba_jll.__init__()
      0.7 ms  micromamba_jll
      0.3 ms  MicroMamba
```

master:
```julia
$ ./julia -e 'using InteractiveUtils; @time_imports using MicroMamba'
      0.8 ms  Printf
     33.2 ms  Dates
      0.5 ms  Scratch
      0.9 ms  TOML
               ┌ 0.0 ms NetworkOptions.__init__()
      4.8 ms  NetworkOptions
               ┌ 0.2 ms MbedTLS_jll.__init__()
      4.7 ms  MbedTLS_jll
               ┌ 0.1 ms LibSSH2_jll.__init__()
      3.9 ms  LibSSH2_jll
               ┌ 0.2 ms LibGit2_jll.__init__()
      5.8 ms  LibGit2_jll
     23.8 ms  LibGit2
     23.3 ms  ArgTools
               ┌ 0.1 ms nghttp2_jll.__init__()
      4.3 ms  nghttp2_jll
               ┌ 0.2 ms LibCURL_jll.__init__()
      4.1 ms  LibCURL_jll
               ┌ 0.0 ms MozillaCACerts_jll.__init__()
      3.4 ms  MozillaCACerts_jll
               ┌ 0.0 ms LibCURL.__init__()
      3.9 ms  LibCURL
               ┌ 0.2 ms Downloads.Curl.__init__()
     51.9 ms  Downloads
      1.3 ms  Tar
               ┌ 0.1 ms p7zip_jll.__init__()
      5.3 ms  p7zip_jll
      0.3 ms  UUIDs
      0.2 ms  Logging
               ┌ 0.0 ms Pkg.__init__()
    865.5 ms  Pkg
      0.3 ms  LazyArtifacts
     25.6 ms  Preferences
      1.1 ms  JLLWrappers
               ┌ 232.7 ms micromamba_jll.__init__() 99.87% compilation time
    234.6 ms  micromamba_jll 99.06% compilation time
      0.4 ms  MicroMamba
```