@topolarity 2024-09-12
The master results are from 945517b - I'm just running on a totally different set of hardware than you (this was a quick measurement taken from my laptop)