@IanButterworth 2024-09-12
Fair enough. Pkg is so slow for you `865.5 ms  Pkg`