@fingolfin 2024-10-25
So I guess this mainly needs some more comments / buy-in from maybe some core people before its worthwhile to pursue turning it into a full-fledged proposal?

@IanButterworth already commented (thanks), would be good to hear what others like e.g. @KristofferC @staticfloat think