@staticfloat 2024-10-25
I agree with this in principle. All that remains are implementation details.